some size_t to uint32_t

+++ src/external/shishua/shishua-avx2.h
@@ -108,7 +108,7 @@ void prng_init(prng_state * restrict s, const uint64_t * seed) {
   s->state[1] = _mm256_set_epi64x(phi[7], phi[6] ^ seed[3], phi[5], phi[4] ^ seed[2]);
   s->state[2] = _mm256_set_epi64x(phi[11], phi[10] ^ seed[3], phi[9], phi[8] ^ seed[2]);
   s->state[3] = _mm256_set_epi64x(phi[15], phi[14] ^ seed[1], phi[13], phi[12] ^ seed[0]);
-  for (size_t i = 0; i < 13; i++) {
+  for (uint32_t i = 0; i < 13; i++) {
     prng_gen(s, buf, 128);
     s->state[0] = s->output[3];
     s->state[1] = s->output[2];
